adicionado o lombok e feito ajustes nos metodos post e update para convido

+++ src/main/java/com/odeivissonsantos/freereunioes/repositorys/ConvidadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.odeivissonsantos.freereunioes.models.Convidado;
 
 @Repository
-public interface ConvidadoRepository extends JpaRepository<Convidado, Integer> {
+public interface ConvidadoRepository extends JpaRepository<Convidado, Long> {
 
 }
